requirements.txt
playwright~=1.49.1
pytest-playwright
pytest~=8.3.4
allure-pytest
allure-python-commons~=2.13.5